+++ R/amr.R
@@ -16,7 +16,7 @@
 # This R package was created for academic research and was publicly    #
 # released in the hope that it will be useful, but it comes WITHOUT    #
 # ANY WARRANTY OR LIABILITY.                                           #
-# Visit our website for more info: https://msberends.gitab.io/AMR.     #
+# Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
 #' The \code{AMR} Package
@@ -64,4 +64,6 @@
 #' \url{https://gitlab.com/msberends/AMR/issues}
 #' @name AMR
 #' @rdname AMR
+#  # prevent NOTE on R >= 3.6
+#' @importFrom microbenchmark microbenchmark
 NULL