+++ R/catalogue_of_life.R
@@ -16,7 +16,7 @@
 # This R package was created for academic research and was publicly    #
 # released in the hope that it will be useful, but it comes WITHOUT    #
 # ANY WARRANTY OR LIABILITY.                                           #
-# Visit our website for more info: https://msberends.gitab.io/AMR.     #
+# Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
 #' The Catalogue of Life
@@ -33,7 +33,7 @@
 #'   \item{All ~55,000 (sub)species from the kingdoms of Archaea, Bacteria and Protozoa}
 #'   \item{All ~3,500 (sub)species from these orders of the kingdom of Fungi: Eurotiales, Onygenales, Pneumocystales, Saccharomycetales, Schizosaccharomycetales and Tremellales. The kingdom of Fungi is a very large taxon with almost 300,000 different (sub)species, of which most are not microbial (but rather macroscopic, like mushrooms). Because of this, not all fungi fit the scope of this package and including everything would tremendously slow down our algorithms too. By only including the aforementioned taxonomic orders, the most relevant fungi are covered (like all species of \emph{Aspergillus}, \emph{Candida}, \emph{Cryptococcus}, \emph{Histplasma}, \emph{Pneumocystis}, \emph{Saccharomyces} and \emph{Trichophyton}).}
 #'   \item{All ~2,000 (sub)species from ~100 other relevant genera, from the kingdoms of Animalia and Plantae (like \emph{Strongyloides} and \emph{Taenia})}
-#'   \item{All ~15,000 previously accepted names of included (sub)species that have been taxonomically renamed}
+#'   \item{All ~21,000 previously accepted names of included (sub)species that have been taxonomically renamed}
 #'   \item{The complete taxonomic tree of all included (sub)species: from kingdom to subspecies}
 #'   \item{The responsible author(s) and year of scientific publication}
 #' }