+++ R/data.R
@@ -16,7 +16,7 @@
 # This R package was created for academic research and was publicly    #
 # released in the hope that it will be useful, but it comes WITHOUT    #
 # ANY WARRANTY OR LIABILITY.                                           #
-# Visit our website for more info: https://msberends.gitab.io/AMR.     #
+# Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
 #' Data set with ~500 antibiotics
@@ -183,14 +183,14 @@ catalogue_of_life <- list(
 #'
 #' A data set containing old (previously valid or accepted) taxonomic names according to the Catalogue of Life. This data set is used internally by \code{\link{as.mo}}.
 #' @inheritSection catalogue_of_life Catalogue of Life
-#' @format A \code{\link{data.frame}} with 16,911 observations and 4 variables:
+#' @format A \code{\link{data.frame}} with 21,342 observations and 4 variables:
 #' \describe{
-#'   \item{\code{col_id}}{Catalogue of Life ID}
-#'   \item{\code{tsn_new}}{New Catalogue of Life ID}
-#'   \item{\code{fullname}}{Old taxonomic name of the microorganism}
+#'   \item{\code{col_id}}{Catalogue of Life ID that was originally given}
+#'   \item{\code{col_id_new}}{New Catalogue of Life ID that responds to an entry in the \code{\link{microorganisms}} data set}
+#'   \item{\code{fullname}}{Old full taxonomic name of the microorganism}
 #'   \item{\code{ref}}{Author(s) and year of concerning scientific publication}
 #' }
-#' @source [3] Catalogue of Life: Annual Checklist (public online database), \url{www.catalogueoflife.org}.
+#' @source Catalogue of Life: Annual Checklist (public online taxonomic database), \url{www.catalogueoflife.org} (check included annual version with \code{\link{catalogue_of_life_version}()}).
 #' @inheritSection AMR Read more on our website!
 #' @seealso \code{\link{as.mo}} \code{\link{mo_property}} \code{\link{microorganisms}}
 "microorganisms.old"
@@ -261,3 +261,23 @@ catalogue_of_life <- list(
 #' }
 #' @inheritSection AMR Read more on our website!
 "WHONET"
+
+# transforms data set to data.frame with only ASCII values, to comply with CRAN policies
+dataset_UTF8_to_ASCII <- function(df) {
+  trans <- function(vect) {
+    iconv(vect, from = "UTF-8", to = "ASCII//TRANSLIT")
+  }
+  df <- as.data.frame(df, stringsAsFactors = FALSE)
+  for (i in 1:NCOL(df)) {
+    col <- df[, i]
+    if (is.factor(col)) {
+      levels(col) <- trans(levels(col))
+    } else if (is.character(col)) {
+      col <- trans(col)
+    } else {
+      col
+    }
+    df[, i] <- col
+  }
+  df
+}