+++ R/deprecated.R
@@ -16,7 +16,7 @@
 # This R package was created for academic research and was publicly    #
 # released in the hope that it will be useful, but it comes WITHOUT    #
 # ANY WARRANTY OR LIABILITY.                                           #
-# Visit our website for more info: https://msberends.gitab.io/AMR.     #
+# Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
 #' Deprecated functions
@@ -45,20 +45,6 @@ ratio <- function(x, ratio) {
     stop('`x` and `ratio` must be of same size.')
   }
   sum(x, na.rm = TRUE) * (ratio / sum(ratio, na.rm = TRUE))
-}
-
-#' @rdname AMR-deprecated
-#' @export
-guess_mo <- function(...) {
-  .Deprecated(new = "as.mo", package = "AMR")
-  as.mo(...)
-}
-
-#' @rdname AMR-deprecated
-#' @export
-guess_atc <- function(...) {
-  .Deprecated(new = "as.atc", package = "AMR")
-  as.atc(...)
 }
 
 #' @rdname AMR-deprecated