+++ R/portion.R
@@ -16,7 +16,7 @@
 # This R package was created for academic research and was publicly    #
 # released in the hope that it will be useful, but it comes WITHOUT    #
 # ANY WARRANTY OR LIABILITY.                                           #
-# Visit our website for more info: https://msberends.gitab.io/AMR.     #
+# Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
 #' Calculate resistance of isolates
@@ -100,14 +100,14 @@
 #'
 #' # Calculate co-resistance between amoxicillin/clav acid and gentamicin,
 #' # so we can see that combination therapy does a lot more than mono therapy:
-#' septic_patients %>% portion_S(amcl)       # S = 67.1%
-#' septic_patients %>% count_all(amcl)       # n = 1576
+#' septic_patients %>% portion_S(amcl)       # S = 71.4%
+#' septic_patients %>% count_all(amcl)       # n = 1879
 #'
 #' septic_patients %>% portion_S(gent)       # S = 74.0%
 #' septic_patients %>% count_all(gent)       # n = 1855
 #'
-#' septic_patients %>% portion_S(amcl, gent) # S = 92.0%
-#' septic_patients %>% count_all(amcl, gent) # n = 1517
+#' septic_patients %>% portion_S(amcl, gent) # S = 92.3%
+#' septic_patients %>% count_all(amcl, gent) # n = 1798
 #'
 #'
 #' septic_patients %>%